adds a "url" state for keeping tarball in sync with local directory
This is being used for dealing with something like the Android SDK
where we need to extract a tarball on disk. This doesn't truly
enforce that the contents of the tarball match what's on disk, but
should be quick and efficient for our needs. You can always
remove the .salt.md5sum file in the path to make salt truly re-sync
this file

example usage:

 /tmp/asdk:
   url.sync_extract:
     - url: http://dl.google.com/android/android-sdk_r21.0.1-linux.tgz
     - md5sum: eaa5a8d76d692d1d027f2bbcee019644